requirements/dev.txt
# Import requirements from base
-r ./base.txt

# Tests
pytest==7.1.2
pytest-django==4.5.2
pytest-lazy-fixture==0.6.3
pytest-mock==3.7.0
coverage==6.3.2

# Code style
flake8==4.0.1
isort==5.10.1
flake8-isort==4.1.1

# Extentions
django_extensions==3.1.5
django-debug-toolbar==3.2.4

# Generate docker-compose
pyyaml==5.4.1
